state: move config-entries table to new pattern
